+++ src/OpenTelemetry.Resources.Process/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Add new process resource attribute `process.creation.time`.
+  ([#4036](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4036))
+
 ## 1.15.1-beta.1
 
 Released 2026-Apr-21